+++ NEWS.md
@@ -1,5 +1,11 @@
 # multiDiff (development version)
 
+# multiDiff xxxx
+
+* Add function `mdd_data_format`, with `plot` and `print` methods
+* Add function `sim_dat_common`
+* Add function `mdd_test_pre_trend_means`
+
 # multiDiff 0.1.3
 
 * Add function `mdd_DD_simple`